+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - CHANGELOG file
+- --which-result option for generate_osm_graph place import
+
+### Changed
+- Moved to osmnx 1.0.1
 
 ## 0.2.0 (2021-02-19)